deduplicate test_duplicated
